+++ apps/operator-ui/src/app/operatorConsolePages/adminPages/adminDisplay.ts
@@ -1,0 +1,6 @@
+export function formatAccess(value: string) {
+  return value
+    .split("_")
+    .map((part) => part.charAt(0).toUpperCase() + part.slice(1))
+    .join(" ");
+}
